requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
google-ads==22.0.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
itsdangerous==2.1.2
requests==2.31.0
PyYAML==6.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.1
